verifying...

+++ lib/screens/root/menu/menus_screen.dart
@@ -1,42 +1,21 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import '../../../models/menu_model.dart';
+import '../../../config/extensions.dart';
 import '../../../providers/lists_provider.dart';
 import '../../../services/menu_service.dart';
 import '../../../widgets/forms/add_edit_menu_widget.dart';
+import '../../../widgets/generic/data_view_widget.dart';
 
 class MenusScreen extends ConsumerWidget {
   const MenusScreen({super.key});
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
-    final lists = ref.watch(getMenusListProvider);
-
-    void showForm(MenuModel? menu) {
-      showModalBottomSheet(
-        context: context,
-        isScrollControlled: false,
-        shape: const RoundedRectangleBorder(
-          borderRadius: BorderRadius.vertical(top: Radius.circular(16)),
-        ),
-        builder: (context) {
-          return Padding(
-            padding: EdgeInsets.only(
-              bottom: MediaQuery.of(context).viewInsets.bottom,
-              left: 16,
-              right: 16,
-              top: 24,
-            ),
-            child: AddEditMenuWidget(menu: menu),
-          );
-        },
-      );
-    }
-
     return Scaffold(
-      body: lists.when(
-        data: (menus) {
+      body: DataViewWidget(
+        provider: getMenusListProvider,
+        dataBuilder: (menus) {
           return menus.isNotEmpty
               ? ListView.builder(
                   itemCount: menus.length,
@@ -51,15 +30,25 @@ class MenusScreen extends ConsumerWidget {
                           children: [
                             IconButton(
                               icon: const Icon(Icons.edit_rounded),
-                              onPressed: () => showForm(menu),
+                              onPressed: () => context.showAppBottomSheet(
+                                child: AddEditMenuWidget(menu: menu),
+                              ),
                             ),
                             IconButton(
                               icon: const Icon(Icons.delete_rounded),
-                              onPressed: () {
-                                ref
-                                    .read(menuServiceProvider)
-                                    .deleteMenu(menu.id);
-                              },
+                              onPressed: () => context
+                                  .showGenericDialog(
+                                    title: 'Delete',
+                                    content:
+                                        'Are you sure, you want to delete the menu?',
+                                  )
+                                  .then((res) {
+                                    if (res == true) {
+                                      ref
+                                          .read(menuServiceProvider)
+                                          .deleteMenu(menu.id);
+                                    }
+                                  }),
                             ),
                           ],
                         ),
@@ -69,12 +58,12 @@ class MenusScreen extends ConsumerWidget {
                 )
               : const Center(child: Text('No menus available.'));
         },
-        error: (error, stackTrace) => Center(child: Text('Error: $error')),
-        loading: () => const Center(child: CircularProgressIndicator()),
       ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: () => showForm(null),
-        child: Icon(Icons.add_rounded),
+      floatingActionButton: context.fabTo(
+        () => context.showAppBottomSheet(
+          isScrollControlled: true,
+          child: AddEditMenuWidget(),
+        ),
       ),
     );
   }
